Move report Query to new folder
